Revert "fix(checker): preserve conformance flow narrowing"
This reverts commit 04951487b40c09d07df34f554a5f6bf9ae96809c.
