+++ crates/tsz-checker/src/checkers/jsx/props/validation.rs
@@ -1482,29 +1482,6 @@ impl<'a> CheckerState<'a> {
         })
     }
 
-    pub(in crate::checkers_domain::jsx) fn jsx_required_props_are_only_iterator_protocol_artifacts(
-        &mut self,
-        props_type: TypeId,
-    ) -> bool {
-        let Some(shape) = self.get_normalized_jsx_required_props_shape(props_type) else {
-            return false;
-        };
-
-        let mut saw_required = false;
-        shape
-            .properties
-            .iter()
-            .filter(|prop| !prop.optional)
-            .all(|prop| {
-                saw_required = true;
-                matches!(
-                    self.ctx.types.resolve_atom_ref(prop.name).as_ref(),
-                    "[Symbol.iterator]" | "__@iterator" | "next"
-                )
-            })
-            && saw_required
-    }
-
     fn jsx_component_member_signature_kinds(&mut self, component_type: TypeId) -> (bool, bool) {
         let component_type = self.normalize_jsx_component_type_for_resolution(component_type);
         let component_type = self.evaluate_type_with_env(component_type);
@@ -1989,5 +1966,30 @@ impl<'a> CheckerState<'a> {
             anchor_idx,
         );
         Some(false)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn jsx_props_target_selection_avoids_anonymous_display_prefix_decision() {
+        let source = include_str!("validation.rs");
+        let formatted_member_call = ["format_type", "(member)"].join("");
+        let starts_with_object = [".starts_with", "('{')"].join("");
+        let inline_forbidden = format!("{formatted_member_call}{starts_with_object}");
+        for forbidden in [
+            inline_forbidden,
+            [
+                "let display = self.format_type(member);",
+                "let is_anonymous = display.starts_with('{');",
+            ]
+            .join("\n"),
+        ] {
+            assert!(
+                !source.contains(&forbidden),
+                "JSX props target selection must use TypeId/query facts, \
+                 not formatted anonymous-object display prefixes: found {forbidden}"
+            );
+        }
     }
 }
